+++ apps/backoffice-v2/src/pages/MerchantMonitoring/components/MerchantMonitoringTable/hooks/useMerchantMonitoringTableLogic/useMerchantMonitoringTableLogic.tsx
@@ -21,15 +21,11 @@ export const useMerchantMonitoringTableLogic = () => {
     children,
   }) => {
     return UPDATEABLE_REPORT_STATUSES.includes(cell.row.original.status) ? (
-      <Link
-        to={`/${locale}/merchant-monitoring/${cell.row.id}`}
-        className={`d-full flex p-1`}
-        onClick={onClick}
-      >
+      <Link to={`/${locale}/merchant-monitoring/${cell.row.id}`} onClick={onClick}>
         {children}
       </Link>
     ) : (
-      children
+      <div className="opacity-50">{children}</div>
     );
   };