+++ apps/backoffice-v2/src/pages/MerchantMonitoringBusinessReport/MerchantMonitoringBusinessReport.page.tsx
@@ -1,53 +1,32 @@
-import dayjs from 'dayjs';
-import { Link } from 'react-router-dom';
-import React, { forwardRef, FunctionComponent } from 'react';
 import { MERCHANT_REPORT_STATUSES_MAP, UPDATEABLE_REPORT_STATUSES } from '@ballerine/common';
-import { ArrowLeft, ChevronLeft, FileQuestion } from 'lucide-react';
 import {
   Dialog,
   DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
   DialogTrigger,
-  DropdownMenu,
-  DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuTrigger,
   Skeleton,
-  TextArea,
   TextWithNAFallback,
 } from '@ballerine/ui';
+import dayjs from 'dayjs';
+import { ArrowLeft, ArrowRightIcon, ChevronLeft, FileQuestion } from 'lucide-react';
+import React, { forwardRef, FunctionComponent } from 'react';
+import { Link } from 'react-router-dom';
 
-import { ctw } from '@/common/utils/ctw/ctw';
-import { Notes } from '@/domains/notes/Notes';
-import { Card } from '@/common/components/atoms/Card/Card';
-import { Form } from '@/common/components/organisms/Form/Form';
-import { Tabs } from '@/common/components/organisms/Tabs/Tabs';
 import { Button } from '@/common/components/atoms/Button/Button';
-import { Select } from '@/common/components/atoms/Select/Select';
-import { CardTitle } from '@/common/components/atoms/Card/Card.Title';
-import { FormItem } from '@/common/components/organisms/Form/Form.Item';
-import { CardFooter } from '@/common/components/atoms/Card/Card.Footer';
-import { TabsList } from '@/common/components/organisms/Tabs/Tabs.List';
-import { CardHeader } from '@/common/components/atoms/Card/Card.Header';
-import { FormField } from '@/common/components/organisms/Form/Form.Field';
-import { FormLabel } from '@/common/components/organisms/Form/Form.Label';
-import { SelectItem } from '@/common/components/atoms/Select/Select.Item';
+import { Card } from '@/common/components/atoms/Card/Card';
 import { CardContent } from '@/common/components/atoms/Card/Card.Content';
-import { SelectValue } from '@/common/components/atoms/Select/Select.Value';
-import { FormControl } from '@/common/components/organisms/Form/Form.Control';
-import { FormMessage } from '@/common/components/organisms/Form/Form.Message';
-import { TabsTrigger } from '@/common/components/organisms/Tabs/Tabs.Trigger';
-import { TabsContent } from '@/common/components/organisms/Tabs/Tabs.Content';
-import { SelectContent } from '@/common/components/atoms/Select/Select.Content';
-import { SelectTrigger } from '@/common/components/atoms/Select/Select.Trigger';
-import { ScrollArea } from '@/common/components/molecules/ScrollArea/ScrollArea';
-import { NotesButton } from '@/common/components/molecules/NotesButton/NotesButton';
-import { SidebarInset, SidebarProvider } from '@/common/components/organisms/Sidebar/Sidebar';
+import { CardFooter } from '@/common/components/atoms/Card/Card.Footer';
+import { CardHeader } from '@/common/components/atoms/Card/Card.Header';
+import { CardTitle } from '@/common/components/atoms/Card/Card.Title';
+import { Separator } from '@/common/components/atoms/Separator/Separator';
+import { BALLERINE_CALENDLY_LINK } from '@/common/constants';
+import { ctw } from '@/common/utils/ctw/ctw';
+import { BusinessReport } from '@/domains/business-reports/components/BusinessReport/BusinessReport';
+import { NotesButton } from '@/domains/notes/NotesButton';
+import { NotesSheet } from '@/domains/notes/NotesSheet';
 import { MerchantMonitoringReportStatus } from '@/pages/MerchantMonitoring/components/MerchantMonitoringReportStatus/MerchantMonitoringReportStatus';
 import { useMerchantMonitoringBusinessReportLogic } from '@/pages/MerchantMonitoringBusinessReport/hooks/useMerchantMonitoringBusinessReportLogic/useMerchantMonitoringBusinessReportLogic';
+import { BusinessReportOptionsDropdown } from './BusinessReportOptionsDropdown';
 
 export const DialogDropdownItem = forwardRef<
   React.ElementRef<typeof DropdownMenuItem>,
@@ -86,20 +65,14 @@ export const MerchantMonitoringBusinessReport: FunctionComponent = () => {
     onNavigateBack,
     websiteWithNoProtocol,
     businessReport,
-    tabs,
-    activeTab,
     notes,
     isNotesOpen,
-    turnOngoingMonitoringOn,
-    isDeboardModalOpen,
-    setIsDeboardModalOpen,
-    isDropdownOpen,
-    setIsDropdownOpen,
-    form,
-    onSubmit,
-    deboardingReasonOptions,
+    setIsNotesOpen,
     isFetchingBusinessReport,
     locale,
+    isDemoAccount,
+    reportRef,
+    ...dropdownProps
   } = useMerchantMonitoringBusinessReportLogic();
 
   // User should never really get in here, unless he manually sets the id in the URL.
@@ -149,231 +122,116 @@ export const MerchantMonitoringBusinessReport: FunctionComponent = () => {
   }
 
   return (
-    <SidebarProvider
-      open={isNotesOpen}
-      style={{
-        '--sidebar-width': '25rem',
-        '--sidebar-width-mobile': '20rem',
-      }}
-    >
-      <SidebarInset>
-        <section className="flex h-full flex-col px-6 pb-6 pt-4">
-          <div className={`flex justify-between`}>
-            <Button
-              variant={'ghost'}
-              onClick={onNavigateBack}
-              className={'mb-6 flex items-center space-x-px pe-3 ps-1 font-semibold'}
-            >
-              <ChevronLeft size={18} /> <span>Back</span>
+    <section className="flex h-full flex-col px-6 pt-4">
+      <div className={`flex justify-between pb-4`}>
+        <Button
+          variant={'ghost'}
+          onClick={onNavigateBack}
+          className={'flex items-center space-x-px pe-3 ps-1 font-semibold'}
+        >
+          <ChevronLeft size={18} /> <span>View All Reports</span>
+        </Button>
+
+        {isDemoAccount ? (
+          <div className="space-x-6 text-sm">
+            <span>Get a guided walkthrough of the report</span>
+            <Button asChild variant="wp-primary" className="justify-start space-x-2" size="sm">
+              <a href={BALLERINE_CALENDLY_LINK} target="_blank" rel="noreferrer">
+                <span>Book a quick call</span>
+                <ArrowRightIcon className="d-4" />
+              </a>
             </Button>
-            <DropdownMenu open={isDropdownOpen} onOpenChange={setIsDropdownOpen} modal={false}>
-              <DropdownMenuTrigger asChild>
-                <Button
-                  variant="outline"
-                  className={
-                    'px-2 py-0 text-xs aria-disabled:pointer-events-none aria-disabled:opacity-50'
-                  }
-                >
-                  Options
-                </Button>
-              </DropdownMenuTrigger>
-
-              <DropdownMenuContent
-                align="end"
-                onEscapeKeyDown={e => {
-                  if (isDeboardModalOpen) {
-                    e.preventDefault();
-                  }
-
-                  setIsDeboardModalOpen(false);
-                }}
-              >
-                {businessReport?.monitoringStatus === true ? (
-                  <DialogDropdownItem
-                    triggerChildren={
-                      <Button variant={'ghost'} className="justify-start">
-                        Turn Monitoring Off
-                      </Button>
-                    }
-                    open={isDeboardModalOpen}
-                    onOpenChange={setIsDeboardModalOpen}
-                  >
-                    <DialogHeader>
-                      <DialogTitle>Confirm Deboarding</DialogTitle>
-                      <DialogDescription>
-                        Are you sure you want to deboard this merchant (turn the monitoring off)?
-                      </DialogDescription>
-                    </DialogHeader>
-
-                    <Form {...form}>
-                      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-                        <FormField
-                          control={form.control}
-                          name="reason"
-                          render={({ field }) => (
-                            <FormItem>
-                              <Select onValueChange={field.onChange} value={field.value}>
-                                <FormLabel>Reason</FormLabel>
-
-                                <FormControl>
-                                  <SelectTrigger className="h-9 w-full border-input p-1 shadow-sm">
-                                    <SelectValue placeholder="Select a reason" />
-                                  </SelectTrigger>
-                                </FormControl>
-                                <FormMessage />
-                                <SelectContent>
-                                  {deboardingReasonOptions?.map((option, index) => {
-                                    return (
-                                      <SelectItem key={index} value={option}>
-                                        {option}
-                                      </SelectItem>
-                                    );
-                                  })}
-                                </SelectContent>
-                              </Select>
-                            </FormItem>
-                          )}
-                        />
-
-                        <FormField
-                          control={form.control}
-                          name="userReason"
-                          render={({ field }) => (
-                            <FormItem>
-                              <FormLabel>Additional details</FormLabel>
-
-                              <FormControl>
-                                <TextArea {...field} />
-                              </FormControl>
-                              <FormMessage />
-                            </FormItem>
-                          )}
-                        />
-
-                        <DialogFooter className="mt-6 flex justify-end space-x-4">
-                          <Button
-                            type="button"
-                            onClick={() => {
-                              setIsDeboardModalOpen(false);
-                            }}
-                            variant="ghost"
-                          >
-                            Cancel
-                          </Button>
-                          <Button type="submit" variant="destructive">
-                            Turn Off
-                          </Button>
-                        </DialogFooter>
-                      </form>
-                    </Form>
-                  </DialogDropdownItem>
-                ) : (
-                  <Button
-                    onClick={() => {
-                      if (!businessReport?.business.id) {
-                        throw new Error('Business ID is missing');
-                      }
-
-                      turnOngoingMonitoringOn(businessReport.business.id, {
-                        onSuccess: () => {
-                          setIsDeboardModalOpen(false);
-                          setIsDropdownOpen(false);
-                        },
-                      });
-                    }}
-                    variant={'ghost'}
-                    className="justify-start"
-                  >
-                    Turn Monitoring On
-                  </Button>
-                )}
-              </DropdownMenuContent>
-            </DropdownMenu>
           </div>
-          {isFetchingBusinessReport ? (
-            <Skeleton className="h-6 w-32" />
-          ) : (
+        ) : (
+          <BusinessReportOptionsDropdown
+            {...dropdownProps}
+            businessReport={businessReport}
+            isDemoAccount={isDemoAccount}
+          />
+        )}
+      </div>
+
+      {/* This ignores parent's padding and covers the whole width. Since we know that padding-x is 6 (1.5rem * 2),
+          we can easily determine negative margin and width required to properly display the separator. */}
+      <Separator className="-ml-6 mb-4 w-[calc(100%+3rem)]" />
+
+      <div ref={reportRef}>
+        {isFetchingBusinessReport ? (
+          <Skeleton className="h-6 w-32" />
+        ) : (
+          <div className="flex items-center justify-between">
             <TextWithNAFallback as={'h2'} className="pb-4 text-2xl font-bold">
               {websiteWithNoProtocol}
             </TextWithNAFallback>
-          )}
-          {isFetchingBusinessReport ? (
-            <Skeleton className="my-6 h-6 w-2/3" />
-          ) : (
-            <div className={`flex items-center space-x-8 pb-4`}>
-              <div className={`flex items-center`}>
-                <span className={`me-4 text-sm leading-6 text-slate-400`}>Status</span>
-                <MerchantMonitoringReportStatus
-                  reportId={businessReport?.id}
-                  status={businessReport?.status}
-                  businessId={businessReport?.business.id}
-                />
-              </div>
-              <div className={`text-sm`}>
-                <span className={`me-2 leading-6 text-slate-400`}>Created at</span>
-                {businessReport?.displayDate &&
-                  dayjs(new Date(businessReport?.displayDate)).format('HH:mm MMM Do, YYYY')}
-              </div>
-              <div className={`flex items-center space-x-2 text-sm`}>
-                <span className={`text-slate-400`}>Monitoring Status</span>
-                <span
-                  className={ctw('select-none rounded-full d-3', {
-                    'bg-success': businessReport?.monitoringStatus,
-                    'bg-slate-400': !businessReport?.monitoringStatus,
-                  })}
-                >
-                  &nbsp;
-                </span>
-              </div>
-              <NotesButton numberOfNotes={notes?.length} />
-            </div>
-          )}
-          <Tabs defaultValue={activeTab} className="w-full" key={activeTab}>
-            <TabsList className={'mb-4'}>
-              {tabs.map(tab => (
-                <TabsTrigger key={tab.value} value={tab.value} asChild>
-                  <Link
-                    to={{
-                      search: `?activeTab=${tab.value}`,
-                    }}
-                  >
-                    {tab.label}
-                  </Link>
-                </TabsTrigger>
-              ))}
-            </TabsList>
-            <ScrollArea orientation={'vertical'} className={'h-[65vh] 2xl:h-[75vh]'}>
-              {isFetchingBusinessReport ? (
-                <>
-                  <Skeleton className="h-6 w-72" />
-                  <Skeleton className="mt-6 h-4 w-40" />
 
-                  <div className="mt-6 flex h-[24rem] w-full flex-nowrap gap-8">
-                    <Skeleton className="w-2/3" />
-                    <Skeleton className="w-1/3" />
-                  </div>
-                  <Skeleton className="mt-6 h-[16rem]" />
-                </>
-              ) : (
-                tabs.map(tab => (
-                  <TabsContent key={tab.value} value={tab.value}>
-                    {tab.content}
-                  </TabsContent>
-                ))
-              )}
-            </ScrollArea>
-          </Tabs>
-        </section>
-      </SidebarInset>
-      <Notes
-        notes={notes ?? []}
-        noteData={{
-          entityId: businessReport?.business.id || '',
-          entityType: `Business`,
-          noteableId: businessReport?.id || '',
-          noteableType: `Report`,
-        }}
-      />
-    </SidebarProvider>
+            {isDemoAccount && (
+              <BusinessReportOptionsDropdown
+                {...dropdownProps}
+                businessReport={businessReport}
+                isDemoAccount={isDemoAccount}
+              />
+            )}
+          </div>
+        )}
+        {isFetchingBusinessReport ? (
+          <Skeleton className="my-6 h-6 w-2/3" />
+        ) : (
+          <div className={`flex items-center space-x-8 pb-4`}>
+            <div className={`flex items-center`}>
+              <span className={`me-4 text-sm leading-6 text-slate-400`}>Status</span>
+              <MerchantMonitoringReportStatus
+                reportId={businessReport?.id}
+                status={businessReport?.status}
+                businessId={businessReport?.business.id}
+              />
+            </div>
+            <div className={`text-sm`}>
+              <span className={`me-2 leading-6 text-slate-400`}>Created at</span>
+              {businessReport?.displayDate &&
+                dayjs(new Date(businessReport?.displayDate)).format('HH:mm MMM Do, YYYY')}
+            </div>
+            <div className={`flex items-center space-x-2 text-sm`}>
+              <span className={`text-slate-400`}>Monitoring Status</span>
+              <span
+                className={ctw('select-none rounded-full d-3', {
+                  'bg-success': businessReport?.monitoringStatus,
+                  'bg-slate-400': !businessReport?.monitoringStatus,
+                })}
+              >
+                &nbsp;
+              </span>
+            </div>
+            <NotesSheet
+              open={isNotesOpen}
+              onOpenChange={setIsNotesOpen}
+              modal={false}
+              notes={notes ?? []}
+              noteData={{
+                entityId: businessReport?.business.id || '',
+                entityType: `Business`,
+                noteableId: businessReport?.id || '',
+                noteableType: `Report`,
+              }}
+            >
+              <NotesButton numberOfNotes={notes?.length} />
+            </NotesSheet>
+          </div>
+        )}
+        {isFetchingBusinessReport || !businessReport ? (
+          <>
+            <Skeleton className="h-6 w-72" />
+            <Skeleton className="mt-6 h-4 w-40" />
+
+            <div className="mt-6 flex h-[24rem] w-full flex-nowrap gap-8">
+              <Skeleton className="w-2/3" />
+              <Skeleton className="w-1/3" />
+            </div>
+            <Skeleton className="mt-6 h-[16rem]" />
+          </>
+        ) : (
+          <BusinessReport report={businessReport} />
+        )}
+      </div>
+    </section>
   );
 };