+++ services/workflows-service/src/data-analytics/data-investigation.service.ts
@@ -1,3 +1,4 @@
+import { AggregateType } from '@/data-analytics/consts';
 import { ALERT_DEFINITIONS } from './../../scripts/alerts/generate-alerts';
 import { SubjectRecord, TExecutionDetails } from '@/alert/types';
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
@@ -148,6 +149,7 @@ export class DataInvestigationService {
       excludePaymentMethods = false,
       projectId,
       amountThreshold,
+      havingAggregate,
     } = options;
 
     return {
@@ -160,7 +162,7 @@ export class DataInvestigationService {
             },
           }
         : {}),
-      ...(amountThreshold
+      ...(amountThreshold && isEmpty(havingAggregate)
         ? {
             transactionBaseAmount: {
               gte: amountThreshold,