+++ examples/report-generation-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/report-generation-example
 
+## 0.2.32
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.80
+
 ## 0.2.31
 
 ### Patch Changes