+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.83
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.7.109
+
+## 1.2.82
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.82
+
+## 1.2.81
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.81
+
+## 1.2.80
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/ui@0.5.80
+  - @ballerine/config@1.1.33
+
 ## 1.2.79
 
 ### Patch Changes