+++ services/websocket-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/websocket-service
 
+## 0.1.33
+
+### Patch Changes
+
+- bump
+
 ## 0.1.32
 
 ### Patch Changes